requirements.txt
langchain
langchain-google-genai
python-dotenv
streamlit
PyPDF2
faiss-cpu
langchain-community